# openapi.yaml
openapi: 3.0.1
info:
  title: Document Management System API
  description: |
    A comprehensive Document Management System REST API with features including:

    - **Document Management**: Create, read, update, and delete documents
    - **Content Management**: Upload and download content with flexible storage options
    - **Versioning**: Major and minor version control for documents
    - **Content Renditions**: Primary and secondary content renditions with transformations
    - **Full-Text Search**: Lucene-powered search with fielded queries
    - **File Store Management**: Manage multiple file storage locations
    - **Content Transformations**: PDF-to-text and other content transformations

    ## Key Features

    ### Document Types
    - ARTICLE, MANUAL, REPORT, SPREADSHEET, PRESENTATION, IMAGE, VIDEO, AUDIO, OTHER

    ### Storage Options
    - Database storage for small files
    - File system storage for large files
    - Automatic migration between storage types

    ### Search Capabilities
    - Simple text search across all fields
    - Fielded search (name, description, keywords, tags, content)
    - Boolean operators (AND, OR)
    - Phrase search, wildcard search
    - Indexes all document metadata and indexable content

    ### Content Renditions
    - Primary renditions (original content)
    - Secondary renditions (transformed content)
    - Automatic text extraction from PDFs
    - Extensible transformer plugin framework
  contact:
...
  license:
...
servers:
...
paths:
  /api/filestores/{id}:
    get:
      tags:
      - File Stores
      operationId: getFileStoreById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStoreDTO'
    put:
      tags:
      - File Stores
      operationId: updateFileStore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileStoreDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStoreDTO'
    delete:
      tags:
      - File Stores
      operationId: deleteFileStore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/filestores/{id}/deactivate:
    put:
      tags:
      - File Stores
      operationId: deactivateFileStore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStoreDTO'
  /api/filestores/{id}/activate:
    put:
      tags:
      - File Stores
      operationId: activateFileStore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStoreDTO'
...
  /api/filestores:
    get:
      tags:
      - File Stores
      operationId: getAllFileStores
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileStoreDTO'
    post:
      tags:
      - File Stores
      operationId: createFileStore
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileStoreDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStoreDTO'
...
  /api/filestores/{id}/space:
    get:
      tags:
      - File Stores
      operationId: getAvailableSpace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SpaceInfoDTO'
  /api/filestores/{id}/space/check:
    get:
      tags:
      - File Stores
      operationId: hasEnoughSpace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: requiredBytes
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
    FileStoreDTO:
      required:
      - name
      - rootPath
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        rootPath:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        contentCount:
          type: integer
          format: int64
...
    SpaceInfoDTO:
      type: object
      properties:
        totalSpace:
          type: integer
          format: int64
        usedSpace:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        name:
          type: string
        availableSpace:
          type: integer
          format: int64
        formattedAvailableSpace:
          type: string
...